+++ lib/just_tooltip.dart
@@ -4,4 +4,5 @@ export 'src/just_tooltip.dart';
 export 'src/just_tooltip_overlay.dart';
 export 'src/just_tooltip_theme.dart';
 export 'src/tooltip_position_utils.dart';
+export 'src/tooltip_registry.dart' show TooltipRegistry;
 export 'src/tooltip_shape_painter.dart';